+++ Basalt.TestField/Program.cs
@@ -8,11 +8,14 @@ using Basalt.Common.Physics;
 using Basalt.Common.Utils;
 using Basalt.Core.Common.Abstractions.Engine;
 using Basalt.Core.Common.Abstractions.Input;
+using Basalt.Core.Common.Abstractions.Sound;
 using Basalt.Raylib.Components;
 using Basalt.Raylib.Graphics;
 using Basalt.Raylib.Input;
+using Basalt.Raylib.Sound;
 using Basalt.Raylib.Utils;
 using Basalt.TestField;
+using Basalt.TestField.Components;
 using Basalt.Types;
 using Raylib_cs;
 using System.Numerics;
@@ -32,14 +35,17 @@ var initParams = new WindowInitParams
 ResourceCache.Instance.LoadShader("lighting", @"resources/shaders/lighting.fs", @"resources/shaders/lighting.vs");
 ResourceCache.Instance.LoadModel("robot", @"resources/robot.glb", "lighting");
 ResourceCache.Instance.LoadTexture("logo", @"resources/logo.png");
+ResourceCache.Instance.LoadAudio("music", @"resources/country.mp3", ResourceCacheExtensions.AudioLoadRequest.Type.Music);
+ResourceCache.Instance.LoadAudio("sfx", @"resources/boom.wav", ResourceCacheExtensions.AudioLoadRequest.Type.Sound);
 var builder = new EngineBuilder();
 
 builder.AddComponent<IGraphicsEngine, RaylibGraphicsEngine>(() => new RaylibGraphicsEngine(initParams) { LightingShaderCacheKey = "lighting" }, true);
 builder.AddComponent<IPhysicsEngine, PhysicsEngine>(true);
 builder.AddComponent<IEventBus, EventBus>();
 builder.AddComponent<IInputSystem, RaylibInputSystem>();
+builder.AddComponent<ISoundSystem, RaylibSoundSystem>();
 
-builder.AddLogger(new ConsoleLogger());
+builder.AddLogger(new ConsoleLogger(Basalt.Core.Common.Types.LogLevel.Info));
 
 var engine = builder.Build();
 
@@ -60,6 +66,14 @@ player.AddComponent(new LightSource(player, "lighting") { Color = Color.Red, Typ
 
 Engine.CreateEntity(player);
 
+var trigger = new Entity();
+trigger.Id = "entity.trigger";
+trigger.Transform.Position = new Vector3(10, 2.5f, 0);
+trigger.AddComponent(new BoxCollider(trigger) { Size = new Vector3(5), IsTrigger = true });
+trigger.AddComponent(new ModelRenderer(trigger) { ModelCacheKey = "cube", Size = new Vector3(5), ColorTint = Color.Blue });
+trigger.AddComponent(new Rigidbody(trigger) { IsKinematic = true });
+trigger.AddComponent(new TestTrigger(trigger));
+Engine.CreateEntity(trigger);
 
 
 TestingUtils.SetupTestingScene(250);