+++ Basalt/Common/Utils/ResourceCache.cs
@@ -37,6 +37,7 @@ namespace Basalt.Common.Utils
 		/// <param name="resource">The resource to cache.</param>
 		public static void CacheResource(string resourceName, object resource)
 		{
+			resourceName = resourceName.ToLower().Trim();
 			lock (padlock)
 			{
 				if (!Instance.resourceCache.ContainsKey(resourceName))
@@ -52,13 +53,14 @@ namespace Basalt.Common.Utils
 		/// <typeparam name="T">The type of the resource.</typeparam>
 		/// <param name="resourceName">The name of the resource.</param>
 		/// <returns>The resource with the specified name, or the default value if not found.</returns>
-		public static T GetResource<T>(string resourceName)
+		public static T? GetResource<T>(string resourceName)
 		{
+			resourceName = resourceName.ToLower().Trim();
 			lock (padlock)
 			{
-				if (Instance.resourceCache.ContainsKey(resourceName))
+				if (Instance.resourceCache.ContainsKey(resourceName) && TryGetResource(resourceName, out T result))
 				{
-					return (T)Instance.resourceCache[resourceName];
+					return result;
 				}
 				else
 				{
@@ -76,12 +78,19 @@ namespace Basalt.Common.Utils
 		/// <returns>true if the resource with the specified name is found; otherwise, false.</returns>
 		public static bool TryGetResource<T>(string resourceName, out T resource)
 		{
+			resourceName = resourceName.ToLower().Trim();
 			lock (padlock)
 			{
 				if (Instance.resourceCache.ContainsKey(resourceName))
 				{
-					resource = (T)Instance.resourceCache[resourceName];
-					return true;
+					var r = Instance.resourceCache[resourceName];
+					if(r is T t)
+					{
+						resource = t;
+						return true;
+					}
+					resource = default;
+					return false;
 				}
 				else
 				{
@@ -98,6 +107,7 @@ namespace Basalt.Common.Utils
 		/// <param name="filePath">The path to the file.</param>
 		public static void LoadResourceFromFile(string resourceName, string filePath)
 		{
+			resourceName = resourceName.ToLower().Trim();
 			lock (padlock)
 			{
 				if (File.Exists(filePath))
@@ -117,6 +127,7 @@ namespace Basalt.Common.Utils
 		/// <param name="loadFunc">The custom load function.</param>
 		public static void LoadResourceFromFile(string resourceName, string filePath, Func<string, object> loadFunc)
 		{
+			resourceName = resourceName.ToLower().Trim();
 			lock (padlock)
 			{
 				if (File.Exists(filePath))
@@ -134,6 +145,7 @@ namespace Basalt.Common.Utils
 		/// <param name="filePath">The path to the file.</param>
 		public static void SaveResourceToFile(string resourceName, string filePath)
 		{
+			resourceName = resourceName.ToLower().Trim();
 			lock (padlock)
 			{
 				if (Instance.resourceCache.ContainsKey(resourceName))
@@ -151,6 +163,7 @@ namespace Basalt.Common.Utils
 		/// <param name="resourceName">The name of the resource.</param>
 		public static void UnloadResource(string resourceName)
 		{
+			resourceName = resourceName.ToLower().Trim();
 			lock (padlock)
 			{
 				if (Instance.resourceCache.ContainsKey(resourceName))
@@ -167,6 +180,7 @@ namespace Basalt.Common.Utils
 		/// <returns>True if there is a loaded resource with that key, false otherwise</returns>
 		public static bool HasResourceKey(string resourceName)
 		{
+			resourceName = resourceName.ToLower().Trim();
 			lock (padlock)
 			{
 				return Instance.resourceCache.ContainsKey(resourceName);